Moved tts directory out of src and updated file references
